fix : 게시물 수정, 삭제 기능 컨트롤러 서비스 병합
